pyproject.toml
[tool.poetry]
name = "django_orm"
version = "0.1.0"
description = ""
authors = ["RomanL91 <dev_roman91@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.4"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
